refactor: add file for validation functions
